requirements.txt
streamlit
pandas
mysql-connector-python
easyocr
pillow
openpyxl